schema file septated
