+++ model/requriedString.js
@@ -1,0 +1,8 @@
+const mongoose = require('mongoose');
+
+const requiredString = {
+    type: String,
+    required: true
+};
+
+module.exports = requiredString;
