+++ model/userSchema.js
@@ -1,34 +1,17 @@
 const mongoose = require('mongoose');
+const requiredString = require('./requriedString');
 
 const userSchema = new mongoose.Schema({
-    name: {
-        type: String,
-        require: true,
-    },
-    email: {
-        type: String,
-        require: true,
-    },
+    name: requiredString,
+    email: requiredString,
     phone: {
         type: Number,
         require: true,
     },
-    villageORcity: {
-        type: String,
-        require: true,
-    },
-    district: {
-        type: String,
-        require: true,
-    },
-    password: {
-        type: String,
-        require: true,
-    },
-    cPassword: {
-        type: String,
-        require: true,
-    },
+    villageORcity: requiredString,
+    district: requiredString,
+    password: requiredString,
+    cPassword: requiredString,
     date: {
         type: Date,
         default: Date.now
